Continued updates to Zenodo archiving
